Info Dialog & Dialog fixes
