+++ src/app/smarthome/lib/dialogs/ConfirmDelete.dart
@@ -1,4 +1,4 @@
-import 'dart:ui';
+import 'package:Smarthome/dialogs/DialogWrapper.dart';
 import 'package:flutter/material.dart';
 
 class ConfirmDeleteDialog extends StatelessWidget {
@@ -12,59 +12,47 @@ class ConfirmDeleteDialog extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BackdropFilter(
-      filter: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
-      child: new SimpleDialog(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(15.0),
+    return DialogWrapper(
+      text: "Bestätigung",
+      children: [
+        Text(
+          "Möchten Sie das Gebäude löschen?",
+          style: TextStyle(fontSize: 15.5),
         ),
-        contentPadding: const EdgeInsets.all(25.0),
-        title: Text(
-          "Bestätigung",
-          style: TextStyle(
-            fontSize: 25.0,
-          ),
-        ),
-        children: [
-          Text(
-            "Möchten Sie das Gebäude löschen?",
-            style: TextStyle(fontSize: 15.5),
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              TextButton(
-                onPressed: () {
-                  this.action();
-                  Navigator.pop(context);
-                },
-                child: Text("Ja"),
-              ),
-              TextButton(
-                onPressed: () => Navigator.pop(context),
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Colors.red,
-                    borderRadius: BorderRadius.circular(10.0),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.end,
+          children: [
+            TextButton(
+              onPressed: () {
+                this.action();
+                Navigator.pop(context);
+              },
+              child: Text("Ja"),
+            ),
+            TextButton(
+              onPressed: () => Navigator.pop(context),
+              child: Container(
+                decoration: BoxDecoration(
+                  color: Colors.red,
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+                child: Padding(
+                  padding: EdgeInsets.symmetric(
+                    vertical: 10.0,
+                    horizontal: 20.0,
                   ),
-                  child: Padding(
-                    padding: EdgeInsets.symmetric(
-                      vertical: 10.0,
-                      horizontal: 20.0,
-                    ),
-                    child: Text(
-                      "Nein",
-                      style: TextStyle(
-                        color: Colors.white,
-                      ),
+                  child: Text(
+                    "Nein",
+                    style: TextStyle(
+                      color: Colors.white,
                     ),
                   ),
                 ),
               ),
-            ],
-          ),
-        ],
-      ),
+            ),
+          ],
+        ),
+      ],
     );
   }
 }
